+++ test/controller/UserControllerTest.java
@@ -13,13 +13,6 @@ import static play.test.Helpers.inMemoryDatabase;
 import static play.test.Helpers.running;
 import static play.test.Helpers.testServer;
 
-import java.io.IOException;
-
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.EntityTransaction;
-import javax.persistence.Persistence;
-
 import org.junit.Test;
 
 import com.fasterxml.jackson.databind.JsonNode;
@@ -29,11 +22,9 @@ import models.TypeUser;
 import play.libs.Json;
 import play.libs.ws.WS;
 import play.libs.ws.WSResponse;
-import play.test.FakeApplication;
-import play.test.WithApplication;
-import util.InitDataLoader;
+import util.AbstractTest;
 
-public class UserControllerTest extends WithApplication {
+public class UserControllerTest extends AbstractTest {
     int        timeout = 4000;
     ObjectNode dataOk;
     ObjectNode dataError1;
@@ -43,16 +34,6 @@ public class UserControllerTest extends WithApplication {
     ObjectNode dataError5;
     ObjectNode dataError6;
     ObjectNode dataError7;
-    
-    public static final String ANSI_RESET = "\u001B[0m";
-    public static final String ANSI_BLACK = "\u001B[30m";
-    public static final String ANSI_RED = "\u001B[31m";
-    public static final String ANSI_GREEN = "\u001B[32m";
-    public static final String ANSI_YELLOW = "\u001B[33m";
-    public static final String ANSI_BLUE = "\u001B[34m";
-    public static final String ANSI_PURPLE = "\u001B[35m";
-    public static final String ANSI_CYAN = "\u001B[36m";
-    public static final String ANSI_WHITE = "\u001B[37m";
 
     public UserControllerTest() throws Exception {
         dataOk = Json.newObject();
@@ -104,36 +85,8 @@ public class UserControllerTest extends WithApplication {
         dataError7.put("type", "OTHER");
     }
 
-    @Override
-    public FakeApplication provideFakeApplication() {
-        return fakeApplication(inMemoryDatabase());
-    }
-
-    public void initializeData() {
-        EntityManagerFactory emf = Persistence.createEntityManagerFactory("memoryPersistenceUnit");
-        EntityManager em = emf.createEntityManager();
-        EntityTransaction trx = em.getTransaction();
-        try {
-
-            // Start the transaction
-            trx.begin();
-            InitDataLoader.load(em, "test/init-data.yml");
-            // Commit and end the transaction
-            trx.commit();
-        } catch (RuntimeException | IOException e) {
-            if (trx != null && trx.isActive()) {
-                trx.rollback();
-            }
-        } finally {
-            // Close the manager
-            em.close();
-            emf.close();
-        }
-    }
-
     @Test
-    public void testFindUserOkRequest() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
+    public void testFindAnUserOkRequest() {
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users/1").get().get(timeout);
@@ -147,13 +100,12 @@ public class UserControllerTest extends WithApplication {
             assertEquals(responseJson.get("username").asText(), "test");
             assertEquals(responseJson.get("email").asText(), "test@testing.dev");
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 
     @Test
-    public void testFindUserNotFound() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
+    public void testFindAnUserNotFound() {
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users/5").get().get(timeout);
@@ -165,13 +117,12 @@ public class UserControllerTest extends WithApplication {
             assertTrue(responseJson.isObject());
             assertEquals(responseJson.get("error").asText(), "Not found 5");
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 
     @Test
     public void testPageUsersOkRequest() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users?page=1&size=1").get().get(timeout);
@@ -188,13 +139,12 @@ public class UserControllerTest extends WithApplication {
             assertNotNull(responseJson.get("link-next"));
             assertNull(responseJson.get("link-prev"));
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 
     @Test
     public void testCreateUserOkRequest() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users").post(dataOk).get(timeout);
@@ -207,13 +157,12 @@ public class UserControllerTest extends WithApplication {
             assertEquals(responseJson.get("id").intValue(), 3);
             assertEquals(responseJson.get("username").asText(), "Yasuo");
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 
     @Test
     public void testCreateUserBadRequest1() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users").post(dataError1).get(timeout);
@@ -225,13 +174,12 @@ public class UserControllerTest extends WithApplication {
             assertTrue(responseJson.isObject());
             assertEquals(responseJson.get("username").get(0).asText(), "This field is required");
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 
     @Test
     public void testCreateUserBadRequest2() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users").post(dataError2).get(timeout);
@@ -243,13 +191,12 @@ public class UserControllerTest extends WithApplication {
             assertTrue(responseJson.isObject());
             assertEquals(responseJson.get("email").get(0).asText(), "This field is required");
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 
     @Test
     public void testCreateUserBadRequest3() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users").post(dataError3).get(timeout);
@@ -261,13 +208,12 @@ public class UserControllerTest extends WithApplication {
             assertTrue(responseJson.isObject());
             assertEquals(responseJson.get("password").get(0).asText(), "This field is required");
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 
     @Test
     public void testCreateUserBadRequest4() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users").post(dataError4).get(timeout);
@@ -279,13 +225,12 @@ public class UserControllerTest extends WithApplication {
             assertTrue(responseJson.isObject());
             assertEquals(responseJson.get("type").get(0).asText(), "This field is required");
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 
     @Test
     public void testCreateUserBadRequest5() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users").post(dataError5).get(timeout);
@@ -297,13 +242,12 @@ public class UserControllerTest extends WithApplication {
             assertTrue(responseJson.isObject());
             assertEquals(responseJson.get("username").get(0).asText(), "This username is already registered");
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 
     @Test
     public void testCreateUserBadRequest6() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users").post(dataError6).get(timeout);
@@ -315,13 +259,12 @@ public class UserControllerTest extends WithApplication {
             assertTrue(responseJson.isObject());
             assertEquals(responseJson.get("email").get(0).asText(), "This e-mail is already registered");
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 
     @Test
     public void testCreateUserBadRequest7() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users").post(dataError7).get(timeout);
@@ -333,13 +276,12 @@ public class UserControllerTest extends WithApplication {
             assertTrue(responseJson.isObject());
             assertEquals(responseJson.get("type").get(0).asText(), "Invalid value");
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 
     @Test
     public void testUpdateUserOkRequest() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users/1").put(dataOk.put("id", 1)).get(timeout);
@@ -352,13 +294,12 @@ public class UserControllerTest extends WithApplication {
             assertEquals(responseJson.get("id").intValue(), 1);
             assertEquals(responseJson.get("username").asText(), "Yasuo");
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 
     @Test
     public void testUpdateUserBadRequest1() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users/1").put(dataError1.put("id", 1)).get(timeout);
@@ -370,13 +311,12 @@ public class UserControllerTest extends WithApplication {
             assertTrue(responseJson.isObject());
             assertEquals(responseJson.get("username").get(0).asText(), "This field is required");
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 
     @Test
     public void testUpdateUserBadRequest2() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users/1").put(dataError2.put("id", 1)).get(timeout);
@@ -388,13 +328,12 @@ public class UserControllerTest extends WithApplication {
             assertTrue(responseJson.isObject());
             assertEquals(responseJson.get("email").get(0).asText(), "This field is required");
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 
     @Test
     public void testUpdateUserNotFound() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users/5").put(dataOk.put("id", 5)).get(timeout);
@@ -406,13 +345,12 @@ public class UserControllerTest extends WithApplication {
             assertTrue(responseJson.isObject());
             assertEquals(responseJson.get("id").get(0).asText(), "This user doesn't exist");
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 
     @Test
     public void testDeleteUserOkRequest() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users/1").delete().get(timeout);
@@ -424,13 +362,12 @@ public class UserControllerTest extends WithApplication {
             assertTrue(responseJson.isObject());
             assertEquals(responseJson.get("msg").asText(), "Deleted 1");
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 
     @Test
     public void testDeleteUserNotFound() {
-        System.out.print(ANSI_YELLOW + "Test Name: " + ANSI_PURPLE + Thread.currentThread().getStackTrace()[1].getMethodName() + ANSI_RESET + "\t\t");
         running(testServer(3333, fakeApplication(inMemoryDatabase())), () -> {
             initializeData();
             WSResponse response = WS.url("http://localhost:3333/users/5").delete().get(timeout);
@@ -442,7 +379,7 @@ public class UserControllerTest extends WithApplication {
             assertTrue(responseJson.isObject());
             assertEquals(responseJson.get("error").asText(), "Not found 5");
 
-            System.out.println("[" + ANSI_GREEN + "success" + ANSI_RESET + "]");
+            successTest();
         });
     }
 }
